requirements.txt
pytesseract
opencv-python
Pillow
google-cloud-translate
googletrans
python-dotenv
flask
django
fastapi
uvicorn
freecurrencyapi